refactor(content): isolate retirement implementation
